requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
boto3==1.35.21
botocore==1.35.21
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.8.7
decorator==5.1.1
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
FlagEmbedding==1.2.11
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.18
llama-index-embeddings-bedrock==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-anthropic==0.2.1
llama-index-llms-bedrock==0.2.2
llama-index-llms-openai==0.2.14
llama-index-multi-modal-llms-openai==0.2.2
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.8
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
orderly-set==5.2.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20240706
pdfquery==0.4.3
peft==0.13.2
pexpect==4.9.0
pgvector==0.3.3
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
roman==4.2
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
striprtf==0.0.26
sympy==1.13.2
tabula-py==2.9.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.2
